[GrB] Extract importer method
